bugfix: 补齐工作流结构校验兼容闭环
