+++ server/apps/opspilot/utils/chat_flow_utils/engine/engine.py
@@ -96,69 +96,6 @@ class ChatFlowEngine(FlowGraphMixin, NodeRunnerMixin, SSEResponderMixin):
         self.max_retry_count = 3
         self.execution_timeout = 300  # 5分钟超时
 
-    @staticmethod
-    def _is_valid_graph_id(value: Any) -> bool:
-        return isinstance(value, str) and bool(value.strip())
-
-    @classmethod
-    def _normalize_flow_structure(cls, flow_json: Any) -> tuple[List[str], Dict[str, List[Dict[str, Any]]]]:
-        """返回结构错误和可安全建图的执行视图，不改写持久化配置。"""
-        if not isinstance(flow_json, dict):
-            return ["流程数据必须是对象"], {"nodes": [], "edges": []}
-
-        errors: List[str] = []
-        raw_nodes = flow_json.get("nodes", [])
-        raw_edges = flow_json.get("edges", [])
-
-        if not isinstance(raw_nodes, list):
-            errors.append("nodes 必须是数组")
-            raw_nodes = []
-        if not isinstance(raw_edges, list):
-            errors.append("edges 必须是数组")
-            raw_edges = []
-
-        nodes = []
-        node_ids = set()
-        for index, node in enumerate(raw_nodes, start=1):
-            if not isinstance(node, dict):
-                errors.append(f"节点 {index} 必须是对象")
-                continue
-            node_id = node.get("id")
-            if not cls._is_valid_graph_id(node_id):
-                errors.append(f"节点 {index} 缺少有效 id")
-                continue
-            if not isinstance(node.get("type"), str):
-                errors.append(f"节点 {index} 的 type 必须是字符串")
-                continue
-            if node_id in node_ids:
-                errors.append(f"节点 {index} 的 id 重复: {node_id}")
-                continue
-            node_ids.add(node_id)
-            nodes.append(node)
-
-        edges = []
-        for index, edge in enumerate(raw_edges, start=1):
-            if not isinstance(edge, dict):
-                errors.append(f"边 {index} 必须是对象")
-                continue
-            source = edge.get("source")
-            target = edge.get("target")
-            if not cls._is_valid_graph_id(source):
-                errors.append(f"边 {index} 缺少有效 source")
-                continue
-            if not cls._is_valid_graph_id(target):
-                errors.append(f"边 {index} 缺少有效 target")
-                continue
-            if source not in node_ids:
-                errors.append(f"边 {index} 的 source 未引用现有节点: {source}")
-                continue
-            if target not in node_ids:
-                errors.append(f"边 {index} 的 target 未引用现有节点: {target}")
-                continue
-            edges.append(edge)
-
-        return errors, {"nodes": nodes, "edges": edges}
-
     def _initialize_variables(self, input_data: Dict[str, Any]):
         """初始化变量管理器
 
@@ -355,14 +292,15 @@ class ChatFlowEngine(FlowGraphMixin, NodeRunnerMixin, SSEResponderMixin):
         entry_type = input_data.get("entry_type", "openai")
         logger.info(f"[SSE-Engine] 开始执行 - flow_id: {self.instance.id}, user_id: {user_id}, entry_type: {entry_type}, 节点数: {len(self.nodes)}")
 
+        validation_errors = self.validate_flow()
+        if validation_errors:
+            error_message = f"流程验证失败: {'; '.join(validation_errors)}"
+            return None, None, user_id, entry_type, session_id, "", False, False, self._create_error_response(error_message)
+
         self._initialize_variables(input_data)
 
         node_id = input_data.get("node_id", "")
         self._record_conversation_history(user_id, input_message, "user", entry_type, node_id, session_id)
-
-        validation_errors = self.validate_flow()
-        if validation_errors:
-            return None, None, user_id, entry_type, session_id, node_id, False, False, self._create_error_response("流程验证失败")
 
         start_node = self._get_start_node()
         start_node_type = start_node.get("type", "") if start_node else None
